+++ functions/Show-DriveView.ps1
@@ -1,25 +1,36 @@
 Function Show-DriveView {
-    [cmdletbinding()]
+    [cmdletbinding(DefaultParameterSetName = '__AllParameterSets')]
+    [Alias('sdv')]
+    [OutputType('PSObject',ParameterSetName='raw')]
+    [OutputType('None',ParameterSetName='__AllParameterSets')]
     Param(
         [Parameter(
             Position = 0,
             ValueFromPipeline,
             ValueFromPipelineByPropertyName,
-            HelpMessage = "Specify the name of a remote computer. You must have admin rights. The default is the localhost."
+            HelpMessage = 'Specify the name of a remote computer. You must have admin rights. The default is the localhost.'
         )]
         [Alias('CN')]
         [ValidateNotNullOrEmpty()]
         [string[]]$ComputerName = $env:ComputerName,
-        [Parameter(HelpMessage = "Specify the grid title")]
+
+        [Parameter(HelpMessage = 'Specify the view title')]
         [ValidateNotNullOrEmpty()]
         [string]$Title = 'Drive Report',
-        [parameter(HelpMessage = 'Specify an alternate credential.')]
+
+        [Parameter(
+            ValueFromPipelineByPropertyName,
+            HelpMessage = 'Specify an alternate credential for the remote computer.'
+        )]
         [ValidateNotNullOrEmpty()]
-        [PSCredential]$Credential
+        [PSCredential]$Credential,
+
+        [Parameter(ParameterSetName = 'raw', HelpMessage = 'Display raw output without formatting.')]
+        [switch]$Raw
     )
     DynamicParam {
         # This will imply PowerShell 7
-            If (Get-Module Microsoft.PowerShell.ConsoleGuiTools -ListAvailable) {
+        If (Get-Module Microsoft.PowerShell.ConsoleGuiTools -ListAvailable) {
 
             $paramDictionary = New-Object -Type System.Management.Automation.RuntimeDefinedParameterDictionary
 
@@ -45,7 +56,7 @@ Function Show-DriveView {
         $PSDefaultParameterValues['_verbose:Command'] = $MyInvocation.MyCommand
         $PSDefaultParameterValues['_verbose:block'] = 'Begin'
         _verbose -message $strings.Starting
-        if ($MyInvocation.CommandOrigin -eq "Runspace") {
+        if ($MyInvocation.CommandOrigin -eq 'Runspace') {
             #Hide this metadata when the command is called from another command
             _verbose -message ($strings.PSVersion -f $PSVersionTable.PSVersion)
             _verbose -message ($strings.UsingHost -f $host.Name)
@@ -66,18 +77,18 @@ Function Show-DriveView {
         }
 
         $splat = @{
-            Title = $Title
+            Title      = $Title
             OutputMode = 'None'
         }
     } #begin
 
     Process {
         $PSDefaultParameterValues['_verbose:block'] = 'Process'
-
+        Write-Information $PSBoundParameters -Tags runtime
         foreach ($computer in $ComputerName) {
             _verbose -message ($strings.QueryComputer -f $Computer.toUpper())
             Try {
-                  #create a temporary CIM session
+                #create a temporary CIM session
                 If ($Credential) {
                     _verbose ($strings.RunAs -f $Credential.UserName)
                     $cs = New-CimSession -ComputerName $Computer -Credential $Credential -ErrorAction Stop
@@ -85,19 +96,30 @@ Function Show-DriveView {
                 else {
                     $cs = New-CimSession -ComputerName $Computer -ErrorAction Stop
                 }
-                $data += Get-CimInstance -class win32_logicaldisk -CimSession $cs -ErrorAction Stop |
-                Select-Object @{Name = 'ComputerName'; Expression = { $_.SystemName } },
-                @{Name = 'Drive'; Expression = { $_.DeviceID } },
-                @{Name = 'Type'; Expression = { $TypeMap.item($_.DriveType -as [int]) } },
-                @{Name = 'SizeGB'; Expression = { [int]($_.Size / 1GB) } },
-                @{Name = 'FreeGB'; Expression = { [math]::Round($_.FreeSpace / 1GB, 2) } },
-                @{Name = 'UsedGB'; Expression = { [math]::round(($_.size - $_.FreeSpace) / 1GB, 2) } },
-                @{Name = 'Free%'; Expression = { [math]::round(($_.FreeSpace / $_.Size) * 100, 2) } },
-                @{Name = 'FreeGraph'; Expression = {
-                    #scale the free space to a range of 0 to 50
-                        [int]$per = (($_.FreeSpace / $_.Size) * 100)/2
-                        '|' * $per }
+                $diskData = Get-CimInstance -class win32_logicaldisk -CimSession $cs -ErrorAction Stop
+
+                If ($Raw) {
+                    $data += $diskData | Select-Object @{Name = 'ComputerName'; Expression = { $_.SystemName } },
+                    DeviceID, @{Name = 'Type'; Expression = { $TypeMap.item($_.DriveType -as [int]) } },
+                    Size,FreeSpace,@{Name = 'UsedSpace'; Expression = { $_.Size - $_.FreeSpace } },
+                    @{Name = 'Free%'; Expression = { [math]::round(($_.FreeSpace / $_.Size) * 100, 2) } }
                 }
+                else {
+                    $data += $diskData | Select-Object @{Name = 'ComputerName'; Expression = { $_.SystemName } },
+                    @{Name = 'Drive'; Expression = { $_.DeviceID } },
+                    @{Name = 'Type'; Expression = { $TypeMap.item($_.DriveType -as [int]) } },
+                    @{Name = 'SizeGB'; Expression = { [int]($_.Size / 1GB) } },
+                    @{Name = 'FreeGB'; Expression = { [math]::Round($_.FreeSpace / 1GB, 2) } },
+                    @{Name = 'UsedGB'; Expression = { [math]::round(($_.size - $_.FreeSpace) / 1GB, 2) } },
+                    @{Name = 'Free%'; Expression = { [math]::round(($_.FreeSpace / $_.Size) * 100, 2) } },
+                    @{Name = 'FreeGraph'; Expression = {
+                            #scale the free space to a range of 0 to 50
+                            [int]$per = (($_.FreeSpace / $_.Size) * 100) / 2
+                            '|' * $per }
+                    }
+                }
+
+                Write-Information $data -Tags data
             } #try
             Catch {
                 Write-Error "[$($Computer.toUpper())] $($_.exception.message)"
@@ -112,14 +134,18 @@ Function Show-DriveView {
     End {
         $PSDefaultParameterValues['_verbose:block'] = 'End'
         $PSDefaultParameterValues['_verbose:Command'] = $MyInvocation.MyCommand
-       # $splat['InputObject'] = $data
-        if ($PSBoundParameters.ContainsKey('ConsoleGridView')) {
-           $data | Out-ConsoleGridView @splat
+        Write-Information $splat -Tags runtime
+        if ($Raw) {
+            $data
         }
         else {
-          $data |  Out-GridView @splat
+            if ($PSBoundParameters.ContainsKey('ConsoleGridView')) {
+                $data | Out-ConsoleGridView @splat
+            }
+            else {
+                $data | Out-GridView @splat
+            }
         }
-
         _verbose $strings.Ending
     } #end